Simplify repertoire components
